computer_use_demo/requirements.txt
streamlit>=1.38.0
google-generativeai>=1.1.0
jsonschema==4.22.0